test(e2e): move Kimi branches to helpers
